backend/requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
fiona==1.10.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
geoalchemy2==0.14.2
geopandas==1.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
lxml==4.9.3
rasterio==1.3.9
psycopg2-binary==2.9.9
